scripts/scrape_front_docs/pyproject.toml
[project]
name = "scrape-front-docs"
version = "0.1.0"
description = "Script to scrape Front Academy docs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "python-dotenv",
    "openai",
    "anthropic",
    "google-generativeai",
    "browser-use"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
backend-path = ["."]
